Add get_steps() to tileray, fix bug
When you drive at a slight angle you're mostly going forward, but
occasionally move 1 tile sideways. In order to predict exactly when that
will happen you need to keep track of how many times the vehicle.face
tileray has been advanced since the last change in direction (in order
to be able to create an identical tileray in the simulation).
The += change here is a bug fix. It was unnoticed because nobody used
the steps value for anything.